chore: Move forms out of modal folder
